Rename file.
Odd git issue?
fixes #27
